web/requirements.txt
fastapi
python-magic
requests
uvicorn
